Allow dragging from flyout with field
We do this by letting the event propagate upwards from the shadow block
- not sure why this propagation was blocked - and start the drag
operation.

We also use blockMouseDown_ even in closeable flyouts to ensure single
clicks can open the editor, while drags move into the workspace. This is
modified to use Blockly.dragMode_ so that Field knows when not to open
the editor because the toolbox is being drag-scrolled.